First commit of project

+++ home.php
@@ -16,6 +16,8 @@ $index = "root";
 
 $filesUser = getFiles($index);
 $foldersUser = getFolder($index);
+$AllFiles = getAllFiles($index);
+$AllFolders = getAllFolders($index);
 
 if (isset($_POST['AddFile'])) {
 		
@@ -33,6 +35,11 @@ if (isset($_POST['CreateFolder'])) {
 	$nameFolder =  $mysqli->real_escape_string($_POST['folder']);
 		
 	createFolder($nameFolder, $index);
+	?>
+	<script defer>
+	console.log('Probando');
+	appendAlert('Your folder has been created succesfully!', 'success')</script>
+<?php
 }
 
 ?>
@@ -50,7 +57,8 @@ if (isset($_POST['CreateFolder'])) {
 	<meta name="og:title" content="<?php echo APPTITLE ?>"/>
 	<meta name="og:image" content="<?php echo SERVERURL ?>img/private_server.png"/>
 	<!--	ICONS PAGE	-->
-	<link id="favicon" rel="shortcut icon" href="<?php echo SERVERURL ?>/img/favicon/1x/favicon.png" type="image/png"/>
+		<link id="favicon" rel="shortcut icon" href="<?php echo SERVERURL ?>favicon.svg" >
+
 	<link rel="apple-touch-icon" sizes="194x194" href="<?php echo SERVERURL ?>img/apple-touch-icon.png" type="image/png"/>
 	<!--    NORMALIZE.CSS v8.0.1    -->
 	<link rel="stylesheet" href="<?php echo SERVERURL ?>css/normalize.css">
@@ -58,11 +66,13 @@ if (isset($_POST['CreateFolder'])) {
 	<link rel="stylesheet" href="<?php echo SERVERURL ?>css/elements.css">
 	<link rel="stylesheet" href="<?php echo SERVERURL ?>css/scrollbar.css">
 	<link rel="stylesheet" href="<?php echo SERVERURL ?>css/home.css">
+	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css">
 	<!--	ICONS fontawesome-free	-->
 	<link rel="stylesheet" href="<?php echo SERVERURL ?>plugins/fontawesome-free/css/all.min.css">
 	<!--    SCRIPT JS    --->
 	<script src="<?php echo SERVERURL ?>js/script.js"></script>
 	<title><?php echo TITLEHOME ?></title>
+	<script defer src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
 </head>
 <body >
 
@@ -97,7 +107,7 @@ if (isset($_POST['CreateFolder'])) {
 	<input type="checkbox" name="btnuserImage" id="btnuserImage">
 	
 	<section class="header">
-		<header><label for="btnnavbar"><i class="fas fa-bars"></i></label><span>&nbsp;Private <strong>Server<i class="fas fa-lock"></i></strong></span></header>
+		<header><label for="btnnavbar"><i class="fas fa-bars"></i></label><span>&nbsp;Records <strong>Buffets<i class="fas fa-lock"></i></strong></span></header>
 		
 		<div class="userImage">
 			<label for="btnuserImage" id="labeluserImage">
@@ -112,14 +122,15 @@ if (isset($_POST['CreateFolder'])) {
 		</div>
 	</section>
 
-	<section class="operations">
+	<?php if ($_SESSION['id_tipo'] != 3){ ?>
+		<section class="operations">
 		<input type="checkbox" name="btnaddFile" id="btnaddFile">
 		<div class="operations-item">
 			<label for="btnaddFile" title="Add File">
 				<i class="fas fa-plus-circle"></i>
 			</label>
 			<div class="inputAddFile">
-				<form enctype="multipart/form-data" method="post" action="<?php echo $_SERVER['PHP_SELF']; ?>">
+				<form enctype="multipart/form-data" method="post" action="<?php echo $_SERVER['PHP_SELF']; ?>" id="fileForm">
 					<label for="btnaddFile">✖</label>
 					<p class="title-login"><i class="fas fa-file"></i>Add File</p>		
 					<input type="file" name="file" id="file" onchange="cambiar()" required>
@@ -143,14 +154,16 @@ if (isset($_POST['CreateFolder'])) {
 			</div>	
 		</div>
 	</section>
+	<?php } ?>
+	
 
 	<section class="files">
-
-		<?php if ($foldersUser != null){ ?>
-			<?php foreach ($foldersUser as $folder ){ ?>
+		<?php if ($_SESSION['id_tipo'] != 2){ ?>
+			<?php if ($AllFolders != null){ ?>
+			<?php foreach ($AllFolders as $folder ){ ?>
 
 				<?php if ($folder['name'] != "root") { ?>
-					<div class="cardfolder">
+					<article class="cardfolder">
 					<a href="<?php echo SERVERURL;?>page.php?cid=<?php echo $folder['cid']?>" class="link">
 						<div class="img">
 							<img src="<?php echo SERVERURL;?>images/folder.png">
@@ -159,16 +172,59 @@ if (isset($_POST['CreateFolder'])) {
 							<span><?php echo subname($folder['name']); ?></span>
 							<div class="files-options">
 								<a href="<?php echo SERVERURL;?>service/downloadFolder.php?cid=<?php echo $folder['cid']?>"><i class="fas fa-download"></i></a>
-								<a href="<?php echo SERVERURL;?>service/share.php?cid=<?php echo $folder['cid']?>"><i class="far fa-share-square"></i></a>
+								<?php if ($_SESSION['id_tipo']!=3) { ?>
+									<a href="<?php echo SERVERURL;?>service/deleteFolder.php?cid=<?php echo $folder['cid']?>"><i class="far fa-trash-alt"></i></a>
+								<?php } ?>
+							</div>
+						</div>
+					</a>
+				</article>	
+
+			<?php }}} ?>
+			<?php if ($filesUser != null){ ?>
+			<?php foreach ($filesUser as $file ){ ?>
+				<div class="cardfile">
+					<a href="<?php echo SERVERURL;?>service/view.php?type=<?php echo $file['extension'] ?>&cid=<?php echo $file['cid']  ?>" class="link" title="<?php echo $file['name']?>">
+						<div class="img">
+							<img src="<?php echo SERVERURL;?>images/<?php echo $file['extension'] ?>.png">
+						</div>
+						<div class="name">
+							<span><?php echo subname($file['name'])  ?></span>
+							<div class="files-options">
+								<a href="<?php echo SERVERURL;?>service/download.php?type=<?php echo $file['extension'] ?>&cid=<?php echo $file['cid']  ?>"><i class="fas fa-download"></i></a>
+								
+								<?php if ($_SESSION['id_tipo']!=3) { ?>
+									<a href="<?php echo SERVERURL;?>service/delete.php?type=<?php echo $file['extension'] ?>&cid=<?php echo $file['cid']  ?>"><i class="far fa-trash-alt"></i></a>
+								<?php } ?>
+							</div>
+						</div>
+					</a>
+				</div>
+			<?php } ?>
+		<?php }} ?>	
+
+		<?php if ($_SESSION['id_tipo'] == 2){ ?>
+			<?php if ($foldersUser != null){ ?>
+			<?php foreach ($foldersUser as $folder ){ ?>
+
+				<?php if ($folder['name'] != "root") { ?>
+					<article class="cardfolder">
+					<a href="<?php echo SERVERURL;?>page.php?cid=<?php echo $folder['cid']?>" class="link">
+						<div class="img">
+							<img src="<?php echo SERVERURL;?>images/folder.png">
+						</div>
+						<div class="name">
+							<span><?php echo subname($folder['name']); ?></span>
+							<div class="files-options">
+								<a href="<?php echo SERVERURL;?>service/downloadFolder.php?cid=<?php echo $folder['cid']?>"><i class="fas fa-download"></i></a>
 								<a href="<?php echo SERVERURL;?>service/deleteFolder.php?cid=<?php echo $folder['cid']?>"><i class="far fa-trash-alt"></i></a>
 							</div>
 						</div>
 					</a>
-				</div>	
+				</article>	
 
 		<?php }}} ?>
-
-		
+		</section>
 
 		<?php if ($filesUser != null){ ?>
 			<?php foreach ($filesUser as $file ){ ?>
@@ -181,22 +237,25 @@ if (isset($_POST['CreateFolder'])) {
 							<span><?php echo subname($file['name'])  ?></span>
 							<div class="files-options">
 								<a href="<?php echo SERVERURL;?>service/download.php?type=<?php echo $file['extension'] ?>&cid=<?php echo $file['cid']  ?>"><i class="fas fa-download"></i></a>
-								<a href="<?php echo SERVERURL;?>service/share.php?type=<?php echo $file['extension'] ?>&cid=<?php echo $file['cid']  ?>"><i class="far fa-share-square"></i></a>
 								<a href="<?php echo SERVERURL;?>service/delete.php?type=<?php echo $file['extension'] ?>&cid=<?php echo $file['cid']  ?>"><i class="far fa-trash-alt"></i></a>
 							</div>
 						</div>
 					</a>
 				</div>
 			<?php } ?>
-		<?php } ?>	
+		<?php }} ?>	
 		
 	</section> 
-	
 	<footer>
 		<a href="<?php echo MYWEB ?>" target="_BLANK">
-			<i>Brayan Joya</i>&nbsp;&copy;
+			<i>Buffets Records</i>&nbsp;&copy;
 		</a>
 	</footer>
+</div>
 
+<div id="liveAlertPlaceholder"></div>
+
+<script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js" integrity="sha384-I7E8VVD/ismYTF4hNIPjVp/Zjvgyol6VFvRkX/vR+Vc4jQkC+hVqc2pM8ODewa9r" crossorigin="anonymous"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.min.js" integrity="sha384-BBtl+eGJRgqQAUMxJ7pMwbEyER4l1g+O15P+16Ep7Q9Q+zqX6gSbd85u4mG4QzX+" crossorigin="anonymous"></script>
 </body>
 </html>
